Moved configuration scripts into config folder. Minor changes
