+++ config/authentication.js
@@ -1,7 +1,7 @@
 /* Middleware Definitions */
 const debug = require('debug')('authentication');
 const LocalStrategy = require('passport-local').Strategy;
-const User = require('./models/user');
+const User = require('../models/user');
 const bcrypt = require('bcryptjs');
 
 module.exports = function (app, passport) {
